+++ src/app/collections/apiDocumentation/apiDocumentation.controller.js
@@ -5,7 +5,7 @@
         .controller('ApiDocumentationController', ApiDocumentationController);
 
     /** @ngInject */
-    function ApiDocumentationController ($compile, $scope) {
+    function ApiDocumentationController ($compile, $scope, SPLIT_PRIMARY, SPLIT_SECONDARY) {
         var vm = this;
 
         vm.apiTransform = apiTransform;
@@ -17,12 +17,12 @@
 
         function activate () {
             vm.columnSet = [
-                { predicate: 'developer', display: 'Developer' },
-                { predicate: 'product', display: 'Product' },
-                { predicate: 'version', display: 'Version' },
-                { predicate: 'chplProductNumber', display: 'CHPL ID', sortDefault: true, isLink: true },
-                { predicate: 'apiDocumentation', display: 'API Documentation', transformFn: vm.apiTransform },
-                { predicate: 'transparencyAttestationUrl', display: 'Mandatory Disclosures URL', transformFn: vm.disclosuresTransform },
+                { predicate: 'developer', display: 'Developer', sortType: 'single' },
+                { predicate: 'product', display: 'Product', sortType: 'single' },
+                { predicate: 'version', display: 'Version', sortType: 'single' },
+                { predicate: 'chplProductNumber', display: 'CHPL ID', sortType: 'single', sortDefault: true, isLink: true },
+                { predicate: 'apiDocumentation', display: 'API Documentation', sortType: 'single', transformFn: vm.apiTransform },
+                { predicate: 'transparencyAttestationUrl', display: 'Mandatory Disclosures URL', sortType: 'single', transformFn: vm.disclosuresTransform },
             ];
         }
 
@@ -32,10 +32,10 @@
             var ret = 'Unknown';
             if (data) {
                 var apis = {};
-                var pairs = data.split('☺');
+                var pairs = data.split(SPLIT_PRIMARY);
                 var key, value;
                 for (var i = 0; i < pairs.length; i++) {
-                    var items = pairs[i].split('☹');
+                    var items = pairs[i].split(SPLIT_SECONDARY);
                     key = items[0];
                     value = items[1];
                     if (value) {