+++ src/app/collections/inactive/inactive.controller.js
@@ -16,14 +16,14 @@
 
         function activate () {
             vm.columnSet = [
-                { predicate: 'edition', display: 'Edition' },
-                { predicate: 'developer', display: 'Developer' },
-                { predicate: 'product', display: 'Product' },
-                { predicate: 'version', display: 'Version' },
-                { predicate: 'decertificationDate', display: 'Inactive As Of', isDate: true },
-                { predicate: 'numMeaningfulUse', display: '# of Known Users', nullDisplay: 'Unknown' },
-                { predicate: 'acb', display: 'ONC-ACB' },
-                { predicate: 'chplProductNumber', display: 'CHPL ID', sortDefault: true, isLink: true },
+                { predicate: 'edition', display: 'Edition', sortType: 'single' },
+                { predicate: 'developer', display: 'Developer', sortType: 'single' },
+                { predicate: 'product', display: 'Product', sortType: 'single' },
+                { predicate: 'version', display: 'Version', sortType: 'single' },
+                { predicate: 'decertificationDate', display: 'Inactive As Of', sortType: 'single', isDate: true },
+                { predicate: 'numMeaningfulUse', display: '# of Known Users', sortType: 'single', nullDisplay: 'Unknown' },
+                { predicate: 'acb', display: 'ONC-ACB', sortType: 'single' },
+                { predicate: 'chplProductNumber', display: 'CHPL ID', sortType: 'single', sortDefault: true, isLink: true },
             ];
             vm.filters = ['acb', 'decertificationDate', 'edition'];
             vm.refineModel = {