+++ RELEASE_NOTES.md
@@ -1,7 +1,7 @@
 # Release Notes
 
-## Version TBD
-_Date TBD_
+## Version 9.2.0
+_17 July 2017_
 
 ### New features
 * Added "Transparency Attestations" for Developers Collection